Update dockerfile to copy package.json before running npm install
We recently packaged codeclimate-fixme as a node module and so
updated the Dockerfile to run simply `npm install`, grabbing the
requisite dependencies from the package.json.

However, we need to make sure that the package.json gets copied
over first. This change fixes the Dockerfile and makes it look
more like the flow in some of our other engines.

E.g.: codeclimate-rubocop
https://github.com/codeclimate/codeclimate-rubocop/blob/master/Dockerfile