Don't try to parse output until finished
I switched to spawn from exec in order to try to stream issues as they
were reported by grep but node will pass chunks of output that may not
terminate with a line break causing a line to straddle two chunks. For
now buffer all of the data and output issues on exit. In another PR
I'll clean this up.